requirements.txt
pandas
matplotlib.pyplot
seaborn
streamlit
babel.numbers